+++ entity/src/main/java/com/wilma/entity/users/RemoteClient.java
@@ -1,4 +1,4 @@
-package com.wilma.entity.forum.users;
+package com.wilma.entity.users;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
