+++ entity/src/main/java/com/wilma/entity/users/Role.java
@@ -1,5 +1,5 @@
 
-package com.wilma.entity.forum.users;
+package com.wilma.entity.users;
 
 import lombok.*;
