+++ repository/src/main/java/com/wilma/repository/RoleRepository.java
@@ -1,6 +1,6 @@
 package com.wilma.repository;
 
-import com.wilma.entity.forum.users.Role;
+import com.wilma.entity.users.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
