+++ repository/src/main/java/com/wilma/repository/UserAccountRepository.java
@@ -1,7 +1,7 @@
 
 package com.wilma.repository;
 
-import com.wilma.entity.forum.users.UserAccount;
+import com.wilma.entity.users.UserAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
