+++ service/src/main/java/com/wilma/service/security/CustomUserDetailsService.java
@@ -1,7 +1,7 @@
 
 package com.wilma.service.security;
 
-import com.wilma.entity.forum.users.UserAccount;
+import com.wilma.entity.users.UserAccount;
 import com.wilma.repository.UserAccountRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
